ui/requirements.txt
PyQt6==6.6.1
websocket-client>=1.7.0